requirements.txt
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
babel==2.17.0
beautifulsoup4==4.14.2
cachetools==6.2.0
certifi==2025.10.5
chardet==5.2.0
charset-normalizer==3.4.3
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.0
feedparser==6.0.12
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.41.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
html5lib==1.1
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
jiter==0.11.0
jusText==3.0.2
langdetect==1.0.9
lxml==5.4.0
lxml_html_clean==0.4.3
openai==2.2.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
readability-lxml==0.8.4.1
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
tenacity==9.1.2
tld==0.13.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
webencodings==0.5.1
websockets==15.0.1
